requirements_debug.txt
-i https://pypi.python.org/simple
backcall==0.1.0
certifi==2018.11.29
decorator==4.3.2
django-debug-toolbar-template-timings==0.9
django-debug-toolbar==1.11
django==2.1.7
ipython-genutils==0.2.0
ipython==7.3.0
jedi==0.13.3
parso==0.3.4
pexpect==4.6.0 ; sys_platform != 'win32'
pickleshare==0.7.5
pipdeptree==0.13.2
pipenv==2018.11.26
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pygments==2.3.1
pytz==2018.9
six==1.12.0
sqlparse==0.2.4
traitlets==4.3.2
virtualenv-clone==0.5.1
virtualenv==16.4.3
wcwidth==0.1.7
